# allow n-dimensional arrays for pet_steady, speedup p_sat calculation