--- a/tests/test_psychrometrics.py
+++ b/tests/test_psychrometrics.py
@@ -1,4 +1,5 @@
 import numpy as np
+import pytest
 
 from pythermalcomfort.psychrometrics import (
     t_dp,
@@ -30,12 +31,12 @@ def test_enthalpy():
 
 def test_psy_ta_rh():
     assert psy_ta_rh(25, 50, p_atm=101325) == PsychrometricValues(
-        p_sat=3169.2,
-        p_vap=1584.6,
-        hr=0.009881547577511219,
+        p_sat=pytest.approx(3169.2, abs=1e-1),
+        p_vap=pytest.approx(1584.6, abs=1e-1),
+        hr=pytest.approx(0.009881547577511219, abs=1e-7),
         t_wb=18.0,
         t_dp=13.8,
-        h=50259.66,
+        h=pytest.approx(50259.79, abs=1e-1),
     )
 
 
@@ -50,8 +51,8 @@ def test_t_o():
 
 
 def test_p_sat():
-    assert (p_sat(tdb=25)) == 3169.2
-    assert (p_sat(tdb=50)) == 12349.9
+    assert pytest.approx(p_sat(tdb=25), abs=1e-1) == 3169.2
+    assert pytest.approx(p_sat(tdb=50), abs=1e-1) == 12349.9
 
 
 def test_t_mrt():