@jkittner 2024-10-13
Ah yeah - your tests are broken. You can't do that. Something changed in the other repo. Also network connection during tests is not always something you can rely on. This will always be flaky...

https://github.com/CenterForTheBuiltEnvironment/pythermalcomfort/blob/496f3799de287737f2ea53cc6a8c900052a29aaa/tests/test_pmv.py#L11

I would simply add the test files to this repo. Or since they are test-specific even hardcode them into the test itself...