@FedericoTartarini 2024-10-14
@jkittner thank you so much for working on this and for submitting this pull request. 

## Pull request
I apologise in advance since I did not clearly specify this in the documentation, however, I would like to ask you to submit your pull request to the `development` branch and not to `master`. Could you please do that?

Feel also free to update the documentation if you have time.

## Tests
There is a group of students who is actively working on pythermalcomfort and we are implementing some big changes on how tests are run, this is the reason why some of the tests may be failing. 

### Network connection for tests
I completely understand your concern of not having to rely on the network to run the tests. I am happy to hear suggestions on how we can fix that. However, we want to move all the validation tables to a new repository which is going to contain the validation tables for `jsthermalcomfort` and `comf`. Currently, we are requesting these data each time we run a test, however, we may want to consider another approach such as using submodules but I am not vary familiar with that.