@jkittner 2024-10-14
okay, I rebased and made the PR against the development branch.

If you want a centralized location for the test data because you're using it for multiple repos, I'd also suggest a submodule which you can pin at a specific commit-sha so it won't break your tests the moment you change something in the data-repo.