@jkittner 2024-10-14
since removing the `round` did give me such a performance gain, I had a look at where else it is used:

`git grep "round(" | grep -v tests | wc -l` returned `95` entries (probably better to use `pytest.approx` in tests)

I am not a huge fan of that tbh - rounding over and over again will make you slowly loose precision. I think it's up to a user to decide when to round (likely only the final result)