@FedericoTartarini 2024-10-14
thank you for creating a pull request to `development`.

## testing
> If you want a centralized location for the test data because you're using it for multiple repos, I'd also suggest a submodule which you can pin at a specific commit-sha so it won't break your tests the moment you change something in the data-repo.

great, the students will not have time to fix this but this is something I can try to fix in the future. I am going to open a new issue so I do not forget

## rounding
Thank you for the suggestion, if you have time we can also try to fix that and perhaps consider removing the round function, we will have, however, to make sure that this is not going to break the tests.